Backend/requirements.txt
accelerate==0.21.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
black==23.9.0
camel-converter==3.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.27.1
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
diffusers==0.19.3
docopt==0.6.2
docutils==0.20.1
elastic-transport==8.4.0
elasticsearch==8.9.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.1.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.7.4
fastapi==0.101.0
ffmpy==0.3.1
filelock==3.12.2
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
fuzzywuzzy==0.18.0
gradio==3.39.0
gradio_client==0.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1691424382338/work
ipython @ file:///croot/ipython_1691532092695/work
jaraco.classes==3.3.0
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jsons==1.6.3
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081219/work
keyring==24.2.0
kiwisolver==1.4.4
lazy_loader==0.3
Levenshtein==0.21.1
linkify-it-py==2.0.2
lit==16.0.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdit-py-plugins==0.3.3
mdurl==0.1.2
meilisearch==0.28.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
nh3==0.2.14
ninja==1.11.1
nltk==3.8.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.1
openai==0.27.8
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
orjson==3.9.4
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.11.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pillow-avif-plugin==1.3.1
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.10.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.0
psutil==5.9.5
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-Levenshtein==0.21.1
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
rank-bm25==0.2.2
rapidfuzz==3.2.0
readme-renderer==42.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.3
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.27.0
supervision==0.13.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==1.8
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.5
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
towhee==1.1.2
towhee.models==1.1.2
tqdm==4.65.2
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.31.0
triton==2.0.0
twine==4.0.2
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
typish==1.9.3
tzdata==2023.3
uc-micro-py==1.0.2
uform==0.4.2
urllib3==1.26.16
uvicorn==0.23.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==11.0.3
yapf==0.40.1
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
